Add draw_test with stb_truetype font rendering
- Add tests/draw_test with font loading via stb_truetype
- Texture atlas creation at 512x512 with pre-baked glyphs
- Pre-calculated vertex buffers for entire strings
- Alpha blending enabled in shader (SRC_ALPHA / ONE_MINUS_SRC_ALPHA)
- Move draw.c to test-specific to fix cross-test build issues
- Delete old G2/text implementation (sources/text/)
- Add draw_set_viewport() for proper screen size handling

+++ backends/gpu/metal/sources/pipeline.m
@@ -402,9 +402,9 @@ void kore_metal_render_pipeline_init(kore_metal_device *device, kore_metal_rende
 		pipe->depth_stencil_state = (__bridge_retained void *)[metal_device newDepthStencilStateWithDescriptor:depth_stencil_descriptor];
 	}
 
-	pipe->primitive_type = convert_primitive_topology(parameters->primitive.topology);
-	pipe->cull_mode = convert_cull_mode(parameters->primitive.cull_mode);
-	pipe->front_face_winding = convert_front_face(parameters->primitive.front_face);
+	pipe->primitive_type = (uint32_t)convert_primitive_topology(parameters->primitive.topology);
+	pipe->cull_mode = (uint32_t)convert_cull_mode(parameters->primitive.cull_mode);
+	pipe->front_face_winding = (uint32_t)convert_front_face(parameters->primitive.front_face);
 }
 
 void kore_metal_render_pipeline_destroy(kore_metal_render_pipeline *pipe) {
